% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmicalc-package.R
\docType{package}
\name{bmicalc-package}
\alias{bmicalc}
\alias{bmicalc-package}
\title{bmicalc: Calculating Weight Percentiles According to the CDC}
\description{
Given sex, age, and weight, this package will return the interpolated percentiles based on the CDC growth curves.
}
\author{
\strong{Maintainer}: Michael DeWitt \email{medewitt@wakehealth.edu} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

}
\keyword{internal}
