% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weight_percentile.R
\name{generate_weight_percentile}
\alias{generate_weight_percentile}
\title{Generate Weight Percentile}
\usage{
generate_weight_percentile(
  sex = c("Male", "Female"),
  age = 15,
  weight = 50,
  verbose = TRUE
)
}
\arguments{
\item{sex}{a string indicating the biological sex of the individual}

\item{age}{a numeric indicating the age in years of the individual}

\item{weight}{a numeric indicating the weight in kilograms of the individual}

\item{verbose}{a logical indicating if messages should be printed}
}
\description{
Generate Weight Percentile
}
