% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_percentiles}
\alias{weight_percentiles}
\title{Weight Percentile Data}
\format{
a data frame with 4608 rows and 6 varables
\describe{
\item{sex_dsc}{string, sex, one of Male or Female}
\item{age_in_month}{numeric, age in months}
\item{metric}{string, description of percentile and weight}
\item{value}{numeric, the weight in kilograms}
\item{percentile}{numeric, percentile in which the weight belongs}
\item{age_in_years}{numeric, age in years}
}
}
\source{
\url{https://www.cdc.gov/growthcharts/html_charts/wtage.htm#males}
}
\usage{
weight_percentiles
}
\description{
A dataset containing the Centers for Disease Control and Prevention's weight based percentiles
}
\keyword{datasets}
