% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmicalc-package.R
\docType{package}
\name{bmicalc-package}
\alias{bmicalc}
\alias{bmicalc-package}
\title{bmicalc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael DeWitt \email{medewitt@wakehealth.edu} (\href{https://orcid.org/0000-0001-8940-1967}{ORCID})

}
\keyword{internal}
